+++ src/gort/recipes/operations.py
@@ -295,6 +295,9 @@ class PreObservingRecipe(BaseRecipe):
     async def recipe(self):
         """Runs the pre-observing sequence."""
 
+        # Run a clean-up first in case there are any issues with the specs.
+        await self.gort.cleanup(readout=False)
+
         tasks = []
 
         tasks.append(
@@ -306,8 +309,6 @@ class PreObservingRecipe(BaseRecipe):
             )
         )
         tasks.append(self.gort.telescopes.park(disable=False, kmirror=False))
-        tasks.append(self.gort.nps.calib.all_off())
-        tasks.append(self.gort.guiders.stop())
         tasks.append(self.gort.ags.reconnect())
         tasks.append(self.gort.specs.expose(flavour="bias"))